--- a/lib/reply.js
+++ b/lib/reply.js
@@ -2,7 +2,7 @@
 
 'use strict'
 
-const pump = require('pump')
+const eos = require('end-of-stream')
 const validation = require('./validation')
 const serialize = validation.serialize
 const statusCodes = require('http').STATUS_CODES
@@ -160,7 +160,7 @@ function onSendEnd (reply, payload) {
   }
 
   if (typeof payload.pipe === 'function') {
-    pump(payload, reply.res, pumpCallback(reply))
+    forward(payload, reply.res, reply)
     return
   }
 
@@ -176,15 +176,34 @@ function onSendEnd (reply, payload) {
   reply.res.end(payload)
 }
 
-function pumpCallback (reply) {
-  return function _pumpCallback (err) {
-    if (!err) return
-    if (reply.res.headersSent) {
-      reply.request.log.error(err, 'response terminated with an error with headers already sent')
-    } else {
-      handleError(reply, err)
+function forward (payload, res, reply) {
+  var sourceOpen = true
+
+  eos(payload, { readable: true, writable: false }, function (err) {
+    sourceOpen = false
+    if (err) {
+      if (res.headersSent) {
+        reply.request.log.error(err, 'response terminated with an error with headers already sent')
+        res.destroy()
+      } else {
+        handleError(reply, err)
+      }
     }
-  }
+    // there is nothing to do if there is not an error
+  })
+
+  eos(res, function (err) {
+    if (err) {
+      if (res.headersSent) {
+        reply.request.log.error(err, 'response terminated with an error with headers already sent')
+      }
+      if (sourceOpen) {
+        payload.destroy()
+      }
+    }
+  })
+
+  payload.pipe(res)
 }
 
 function handleError (reply, error, cb) {