--- a/lib/reply.js
+++ b/lib/reply.js
@@ -198,7 +198,13 @@ function forward (payload, res, reply) {
         reply.request.log.error(err, 'response terminated with an error with headers already sent')
       }
       if (sourceOpen) {
-        payload.destroy()
+        if (payload.destroy) {
+          payload.destroy()
+        } else if (typeof payload.close === 'function') {
+          payload.close(noop)
+        } else if (typeof payload.abort === 'function') {
+          payload.abort()
+        }
       }
     }
   })
@@ -274,6 +280,8 @@ function buildReply (R) {
   _Reply.prototype = new R()
   return _Reply
 }
+
+function noop () {}
 
 module.exports = Reply
 module.exports.buildReply = buildReply