--- a/lib/reply.js
+++ b/lib/reply.js
@@ -160,7 +160,7 @@ function onSendEnd (reply, payload) {
   }
 
   if (typeof payload.pipe === 'function') {
-    forward(payload, reply.res, reply)
+    sendStream(payload, reply.res, reply)
     return
   }
 
@@ -176,14 +176,14 @@ function onSendEnd (reply, payload) {
   reply.res.end(payload)
 }
 
-function forward (payload, res, reply) {
+function sendStream (payload, res, reply) {
   var sourceOpen = true
 
   eos(payload, { readable: true, writable: false }, function (err) {
     sourceOpen = false
     if (err) {
       if (res.headersSent) {
-        reply.request.log.error(err, 'response terminated with an error with headers already sent')
+        res.log.error(err, 'response terminated with an error with headers already sent')
         res.destroy()
       } else {
         handleError(reply, err)
@@ -195,7 +195,7 @@ function forward (payload, res, reply) {
   eos(res, function (err) {
     if (err) {
       if (res.headersSent) {
-        reply.request.log.error(err, 'response terminated with an error with headers already sent')
+        res.log.error(err, 'response terminated with an error with headers already sent')
       }
       if (sourceOpen) {
         if (payload.destroy) {