--- a/package.json
+++ b/package.json
@@ -92,6 +92,7 @@
     "tap": "^11.0.1",
     "then-sleep": "^1.0.1",
     "typescript": "^2.6.2",
+    "pump": "^2.0.1",
     "x-xss-protection": "^1.0.0"
   },
   "dependencies": {
@@ -99,13 +100,13 @@
     "abstract-logging": "^1.0.0",
     "ajv": "^6.1.0",
     "avvio": "^5.0.1",
+    "end-of-stream": "^1.4.1",
     "fast-json-stringify": "^0.17.0",
     "find-my-way": "^1.10.0",
     "flatstr": "^1.0.5",
     "light-my-request": "^2.0.1",
     "middie": "^3.1.0",
-    "pino": "^4.10.3",
-    "pump": "^2.0.1"
+    "pino": "^4.10.3"
   },
   "greenkeeper": {
     "ignore": [