I've improved the stream support to not close the destination stream directly if the payload errors, but rather pass it through our error pipeline.

Thanks to this, we can simplify https://github.com/fastify/fastify-static/blob/master/index.js#L34-L48, as the sendStream could be passed in directly

#### Checklist

- [x] run `npm run test` and `npm run benchmark`
- [x] tests and/or benchmarks are included
- [ ] documentation is changed or added
- [x] commit message and code follows [Code of conduct](https://github.com/fastify/fastify/blob/master/CODE_OF_CONDUCT.md)