@allevo 2018-01-30
@mcollina Can you run some benchmarks?