@mcollina 2018-01-30
## Benchmarks

### master

```
$ autocannon -c 100 -d 5 localhost:3000/stream
Running 5s test @ http://localhost:3000/stream
100 connections

Stat         Avg     Stdev  Max
Latency (ms) 24.63   7.17   95.98
Req/Sec      3979.2  296.69 4225
Bytes/Sec    1.67 MB 137 kB 1.78 MB

20k requests in 5s, 8.4 MB read
```

### this branch

```
$ autocannon -c 100 -d 5 localhost:3000/stream
Running 5s test @ http://localhost:3000/stream
100 connections

Stat         Avg     Stdev  Max
Latency (ms) 24.15   6.91   77.37
Req/Sec      4057    312.86 4482
Bytes/Sec    1.71 MB 114 kB 1.89 MB

20k requests in 5s, 8.56 MB read
```